Add logic to handle messages, clear frames, hide/reveal cursor, open/close alternate screen buffer
Messages: messages sent from the reporter as a MessageNotification,
including errors outside of examples and filter announcements,
are stored and displayed at the end of the run.

Hide/reveal Cursor: The cursor is hidden during initialization and
rendered visible again on exit.

Clear Frame: the clear_frame method clears content from the screen
by reseting the cursor to home position [0,0] and clearing the
screen from the cursor downward. This is to prevent the creation
of a scrollable trail of old frames.

Open/Close Buffer: An alternate screen buffer is opened on
initialization and closed on exit.